+++ lib/hanami/routing.rb
@@ -184,6 +184,7 @@ module Hanami
     end
 
     require "hanami/routing/endpoint"
+    require "hanami/routing/endpoint/resolver"
     require "hanami/routing/prefix"
     require "hanami/routing/scope"
     require "hanami/routing/resource"
